Improve build
Create two separate entry points for the libraries, making it possible for
the browser to load both at the same time, improving page speed.

Expose alt and, at the same time, remove SDK responsability to expose an
**util** module of Alt. If the app wants to require this module, just add
alt to the package dependencies and require just the util module,
no extra overhead it's added to page.